Refactor initialisation code for common and Neo4j
