+++ db/neo4j/initialise/Neo4jInitialiser.java
@@ -1,29 +1,27 @@
 package grakn.simulation.db.neo4j.initialise;
 
-import grakn.simulation.db.common.initialise.Initialiser;
 import grakn.simulation.db.common.yaml_tool.YAMLException;
 import grakn.simulation.db.common.yaml_tool.YAMLLoader;
 import grakn.simulation.db.neo4j.yaml_tool.Neo4jYAMLLoader;
+import org.neo4j.driver.Session;
 
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Map;
 
-public class Neo4jInitialiser extends Initialiser {
+public class Neo4jInitialiser {
 
-    public Neo4jInitialiser(Map<String, Path> files) {
-        super(files);
+    private final Session session;
+    private Map<String, Path> files;
+
+    public Neo4jInitialiser(Session session, Map<String, Path> files) {
+        this.session = session;
+        this.files = files;
     }
 
-    protected void initialiseData(DriverWrapper.Session session) throws IOException, YAMLException {
+    public void initialise() throws IOException, YAMLException {
         YAMLLoader loader = new Neo4jYAMLLoader(session, files);
         loader.loadFile(files.get("neo4j_data.yml").toFile());
-    }
-
-    @Override
-    public void initialise(DriverWrapper driverWrapper, String databaseName) throws IOException, YAMLException {
-        DriverWrapper.Session session = driverWrapper.session(databaseName);
-        initialiseData(session);
         // TODO Add key constraints
     }
 }
